+++ internal/httpapi/assets/_app/immutable/chunks/B1WME1IP.js
@@ -1,1 +1,0 @@
-import"./Cj8fpFot.js";import{s as t,r as p}from"./Bt_S5hGO.js";import{I as a}from"./CcI4j89e.js";var n=new Set(["$$slots","$$events","$$legacy"]);function l(s,e){let o=p(e,n);const r=[["path",{d:"M5 12h14"}],["path",{d:"M12 5v14"}]];a(s,t({name:"plus"},()=>o,{get iconNode(){return r}}))}export{l as P};