+++ internal/httpapi/assets/_app/immutable/chunks/BkZvBVYs.js
@@ -1,0 +1,1 @@
+import"./WSFfxB3w.js";import{s,r as p}from"./ThbG16I7.js";import{I as n}from"./C1nGsSwf.js";var a=new Set(["$$slots","$$events","$$legacy"]);function i(e,o){let r=p(o,a);const t=[["path",{d:"M18 6 6 18"}],["path",{d:"m6 6 12 12"}]];n(e,s({name:"x"},()=>r,{get iconNode(){return t}}))}export{i as X};