+++ internal/httpapi/assets/_app/immutable/chunks/C0w6ftS6.js
@@ -1,1 +1,0 @@
-import{v as t,w as i,q as c,x as f}from"./DHWLhpiP.js";import{B as y}from"./Bt_S5hGO.js";const l=Symbol("NaN");function h(r,e,n){c&&f();var o=new y(r),s=!t();i(()=>{var a=e();a!==a&&(a=l),s&&a!==null&&typeof a=="object"&&(a={}),o.ensure(a,n)})}export{h as k};