+++ internal/httpapi/assets/_app/immutable/chunks/Cx5fzKuy.js
@@ -1,0 +1,1 @@
+import"./WSFfxB3w.js";import{s,r as n}from"./ThbG16I7.js";import{I as p}from"./C1nGsSwf.js";var a=new Set(["$$slots","$$events","$$legacy"]);function $(r,e){let o=n(e,a);const t=[["path",{d:"m9 18 6-6-6-6"}]];p(r,s({name:"chevron-right"},()=>o,{get iconNode(){return t}}))}export{$ as C};